Refactor Numba types into a subpackage
