+++ numba/types/abstract.py
@@ -6,8 +6,8 @@ import weakref
 
 import numpy
 
-from .six import add_metaclass
-from .utils import cached_property
+from ..six import add_metaclass
+from ..utils import cached_property
 
 
 # Types are added to a global registry (_typecache) in order to assign
@@ -153,7 +153,7 @@ class Type(object):
     # usable as a function signature).
 
     def __call__(self, *args):
-        from .typing import signature
+        from ..typing import signature
         if len(args) == 1 and not isinstance(args[0], Type):
             return self.cast_python_value(args[0])
         return signature(self, # return_type
@@ -163,7 +163,7 @@ class Type(object):
         """
         Return an array of this type.
         """
-        from .types import Array
+        from . import Array
         ndim, layout = self._determine_array_spec(args)
         return Array(dtype=self, ndim=ndim, layout=layout)
 
@@ -216,7 +216,7 @@ class Number(Hashable):
         """
         Unify the two number types using Numpy's rules.
         """
-        from . import numpy_support
+        from .. import numpy_support
         if isinstance(other, Number):
             # XXX: this can produce unsafe conversions,
             # e.g. would unify {int64, uint64} to float64
